:star: [UR-92] Extract data- prop test into a standalone file
